Initial website
This is my first to-do list webapp, coded in JavaScript, HTML, and CSS with Bootstrap and Animate.css. It is not 100% perfect, but it is functional. Items can be added, selected, and deleted, with animations for every action. Supports clicking the add button as well as pressing the enter key.